more clean-up functions

+++ src/services/Cleaner.ts
@@ -1,8 +1,10 @@
 import * as fs from "fs";
+import { isEmpty } from "lodash";
 import logger from "../config/Logger";
+import { path } from "./VideoLinkedProductLoader";
 
 const cleanUp = async (paths: Array<string>) => {
-    if (paths) {
+    if (paths && !isEmpty(paths)) {
         const promises = paths.map(p =>
             fs.promises
                 .unlink(p)
@@ -14,6 +16,17 @@ const cleanUp = async (paths: Array<string>) => {
         return Promise.all(promises);
     } else {
         return Promise.resolve();
+    }
+}
+
+export const deleteExistingOutFile = () => {
+    try {
+        if (fs.existsSync(path)) {
+            fs.unlinkSync(path);
+        }
+    } catch (error) {
+        logger.error(`failed to clean up before starting, ${JSON.stringify(error)}`);
+        process.exit(2)
     }
 }
